dom/xhr/tests/test_sync_xhr_unload.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1307122</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="common_temporaryFileBlob.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <script type="application/javascript">

SimpleTest.waitForExplicitFinish();

info("Creating the iframe...");
var ifr = document.createElement('iframe');

ifr.addEventListener("load", function ifr_load1() {
  info("Iframe loaded");

  ifr.removeEventListener("load", ifr_load1);
  ifr.src = "empty.html";

  ifr.addEventListener("load", function ifr_load2() {
    ok(true, "Test passed");
    SimpleTest.finish();
  });

});

ifr.src = "iframe_sync_xhr_unload.html";
document.body.appendChild(ifr);

  </script>
</body>
</html>
